requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
pandas==2.3.2
faiss-cpu==1.7.4
langchain==0.3.27
langchain-community==0.3.30
langchain-openai==0.3.33
sentence-transformers==5.1.1
python-dotenv==1.1.1
numpy==1.26.4
langchain_huggingface==0.3.1
langgraph==0.6.7
langdetect==1.0.9